+++ DC1AP/Threads/MiracleChestMgmt.cs
@@ -1,5 +1,6 @@
 ﻿using Archipelago.Core.Util;
 using DC1AP.Constants;
+using DC1AP.Locations;
 using DC1AP.Mem;
 using DC1AP.Utils;
 using System;
@@ -7,7 +8,7 @@ using System.Collections.Generic;
 using System.Threading;
 
 
-namespace DC1AP.Items
+namespace DC1AP.Threads
 {
     internal class MiracleChestMgmt
     {
@@ -167,7 +168,5 @@ namespace DC1AP.Items
                 Memory.Write(PricklyValueAddr, -1);
             }
         }
-
-
     }
 }
